http_apis/requirements.txt
nltk
pyexcel_ods
textract==1.6.3
beautifulsoup4==4.8.0
pypptx
odfpy
PyPDF2==1.26.0
striprtf
numpy
docx2txt
inscriptis
pandas
certifi
chardet
requests
idna
urllib3
pycryptodome
pyOpenSSL
tldextract
graphene>=2.0